Move the testing functionality into a separate file.
